refactor: schema merge to give more control over field merging
