+++ internal_types/Cargo.toml
@@ -8,6 +8,7 @@ readme = "README.md"
 
 [dependencies]
 arrow = { version = "4.0", features = ["prettyprint"] }
+hashbrown = "0.11"
 snafu = "0.6"
 observability_deps = { path = "../observability_deps" }
