+++ internal_types/src/schema.rs
@@ -1,5 +1,5 @@
 //! This module contains the schema definiton for IOx
-use snafu::{ResultExt, Snafu};
+use snafu::Snafu;
 use std::{
     collections::{BTreeSet, HashMap},
     convert::{TryFrom, TryInto},
@@ -33,6 +33,7 @@ pub fn TIME_DATA_TYPE() -> ArrowDataType {
 }
 
 pub mod builder;
+pub mod merge;
 
 /// Database schema creation / validation errors.
 #[derive(Debug, Snafu)]
@@ -50,62 +51,8 @@ pub enum Error {
         actual_type: ArrowDataType,
     },
 
-    #[snafu(display(
-    "Schema Merge Error: Incompatible measurement names. Existing measurement name '{}', new measurement name '{}'",
-    existing_measurement, new_measurement
-    ))]
-    TryMergeDifferentMeasurementNames {
-        existing_measurement: String,
-        new_measurement: String,
-    },
-
-    #[snafu(display(
-        "Schema Merge Error: Incompatible column type for '{}'. Existing type {:?}, new type {:?}",
-        field_name,
-        influx_column_type,
-        existing_influx_column_type
-    ))]
-    TryMergeBadColumnType {
-        field_name: String,
-        existing_influx_column_type: Option<InfluxColumnType>,
-        influx_column_type: Option<InfluxColumnType>,
-    },
-
-    #[snafu(display(
-        "Schema Merge Error: Incompatible data type for '{}'. Existing type {:?}, new type {:?}",
-        field_name,
-        existing_data_type,
-        new_data_type
-    ))]
-    TryMergeBadArrowType {
-        field_name: String,
-        existing_data_type: ArrowDataType,
-        new_data_type: ArrowDataType,
-    },
-
-    #[snafu(display(
-        "Schema Merge Error: Incompatible nullability for '{}'. Existing field {}, new field {}",
-        field_name, nullable_to_str(*existing_nullability), nullable_to_str(*new_nullability)
-    ))]
-    TryMergeBadNullability {
-        field_name: String,
-        existing_nullability: bool,
-        new_nullability: bool,
-    },
-
-    #[snafu(display("Schema Merge: Error merging underlying schema: {}", source))]
-    MergingSchemas { source: arrow::error::ArrowError },
-
     #[snafu(display("Column not found '{}'", column_name))]
     ColumnNotFound { column_name: String },
-}
-
-fn nullable_to_str(nullability: bool) -> &'static str {
-    if nullability {
-        "can be null"
-    } else {
-        "can not be null"
-    }
 }
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
@@ -303,110 +250,7 @@ impl Schema {
         })
     }
 
-    /// Merges any new columns from new_schema, consuming self. If the
-    /// column already exists, self is unchanged. If the column
-    /// definition conflicts with a prior definition, an error is
-    /// returned.
-    pub fn try_merge(self, other: Self) -> Result<Self> {
-        // Optimize for the common case of the same schema
-        let mut need_merge = false;
-
-        // Do our own pre-checks of the new fields to make nicer error messages
-        if let (Some(existing_measurement), Some(new_measurement)) =
-            (self.measurement(), other.measurement())
-        {
-            if existing_measurement != new_measurement {
-                return TryMergeDifferentMeasurementNames {
-                    existing_measurement,
-                    new_measurement,
-                }
-                .fail();
-            }
-        }
-
-        // if one side has a measurement and the other doesn't need to merge
-        if self.measurement() != other.measurement() {
-            need_merge = true;
-        }
-
-        other
-            .iter()
-            .filter_map(|(influx_column_type, field)| {
-                if let Some(idx) = self.find_index_of(field.name()) {
-                    let (existing_influx_column_type, existing_field) = self.field(idx);
-                    Some((
-                        existing_influx_column_type,
-                        existing_field,
-                        influx_column_type,
-                        field,
-                    ))
-                } else {
-                    // new field
-                    need_merge = true;
-                    None
-                }
-            })
-            .try_for_each(
-                |(existing_influx_column_type, existing_field, influx_column_type, field)| {
-                    let field_name = field.name();
-
-                    // for now, insist the types are exactly the same
-                    // (e.g. None and Some(..) don't match). We could
-                    // consider relaxing this constraint
-                    if existing_influx_column_type != influx_column_type {
-                        TryMergeBadColumnType {
-                            field_name,
-                            existing_influx_column_type,
-                            influx_column_type,
-                        }
-                        .fail()
-                    } else if field.data_type() != existing_field.data_type() {
-                        TryMergeBadArrowType {
-                            field_name,
-                            existing_data_type: existing_field.data_type().clone(),
-                            new_data_type: field.data_type().clone(),
-                        }
-                        .fail()
-                    } else if field.is_nullable() != existing_field.is_nullable() {
-                        TryMergeBadNullability {
-                            field_name,
-                            existing_nullability: existing_field.is_nullable(),
-                            new_nullability: field.is_nullable(),
-                        }
-                        .fail()
-                    } else {
-                        Ok(())
-                    }
-                },
-            )?;
-
-        let new_self = if need_merge {
-            // Delegate the rest of the actual work to arrow schema
-            let new_schema = ArrowSchema::try_merge(vec![
-                self.unwrap_to_inner_owned(),
-                other.unwrap_to_inner_owned(),
-            ])
-            .context(MergingSchemas)?;
-            Self {
-                inner: Arc::new(new_schema),
-            }
-        } else {
-            self
-        };
-
-        Ok(new_self)
-    }
-
-    fn unwrap_to_inner_owned(self) -> ArrowSchema {
-        // try and avoid a clone if possible, but it might be required if the Arc is
-        // shared
-        match Arc::try_unwrap(self.inner) {
-            Ok(schema) => schema,
-            Err(schema_arc) => schema_arc.as_ref().clone(),
-        }
-    }
-
-    /// Resort order of our columns lexographically by name
+    /// Resort order of our columns lexicographically by name
     pub fn sort_fields_by_name(self) -> Self {
         // pairs of (orig_index, field_ref)
         let mut sorted_fields: Vec<(usize, &ArrowField)> =
@@ -995,191 +839,6 @@ mod test {
         assert_eq!(iter.next().unwrap().name(), "time");
         assert_eq!(iter.next(), None);
         assert_eq!(iter.next(), None);
-    }
-
-    #[test]
-    fn test_merge_compatible_schema() {
-        let schema1 = SchemaBuilder::new()
-            .tag("the_tag")
-            .influx_field("int_field", Integer)
-            .build()
-            .unwrap();
-
-        // has some of the same and some new, different fields
-        let schema2 = SchemaBuilder::new()
-            .measurement("my_measurement")
-            .tag("the_other_tag")
-            .influx_field("int_field", Integer)
-            .influx_field("another_field", Integer)
-            .build()
-            .unwrap();
-
-        let merged_schema = schema1.try_merge(schema2).unwrap();
-
-        let expected_schema = SchemaBuilder::new()
-            .measurement("my_measurement")
-            .tag("the_tag")
-            .influx_field("int_field", Integer)
-            .tag("the_other_tag")
-            .influx_field("another_field", Integer)
-            .build()
-            .unwrap();
-
-        assert_eq!(
-            expected_schema, merged_schema,
-            "\nExpected:\n{:#?}\nActual:\n{:#?}",
-            expected_schema, merged_schema
-        );
-    }
-
-    #[test]
-    fn test_merge_compatible_schema_no_names() {
-        let schema1 = SchemaBuilder::new().tag("the_tag").build().unwrap();
-
-        // has some of the same and some new, different fields
-        let schema2 = SchemaBuilder::new().tag("the_other_tag").build().unwrap();
-
-        // ensure the merge is not optimized away
-        let merged_schema = schema1.try_merge(schema2).unwrap();
-
-        let expected_schema = SchemaBuilder::new()
-            .tag("the_tag")
-            .tag("the_other_tag")
-            .build()
-            .unwrap();
-
-        assert_eq!(
-            expected_schema, merged_schema,
-            "\nExpected:\n{:#?}\nActual:\n{:#?}",
-            expected_schema, merged_schema
-        );
-    }
-
-    #[test]
-    fn test_merge_compatible_schema_only_measurement() {
-        let schema1 = SchemaBuilder::new()
-            .tag("the_tag")
-            .measurement("the_measurement")
-            .build()
-            .unwrap();
-
-        // schema has same fields but not measurement name
-        let schema2 = SchemaBuilder::new().tag("the_tag").build().unwrap();
-
-        // ensure the merge is not optimized away
-        let merged_schema = schema1.try_merge(schema2).unwrap();
-
-        let expected_schema = SchemaBuilder::new()
-            .tag("the_tag")
-            .measurement("the_measurement")
-            .build()
-            .unwrap();
-
-        assert_eq!(
-            expected_schema, merged_schema,
-            "\nExpected:\n{:#?}\nActual:\n{:#?}",
-            expected_schema, merged_schema
-        );
-    }
-
-    #[test]
-    fn test_merge_measurement_names() {
-        let schema1 = SchemaBuilder::new().tag("the_tag").build().unwrap();
-
-        // has some of the same and some different fields
-        let schema2 = SchemaBuilder::new()
-            .measurement("my_measurement")
-            .build()
-            .unwrap();
-
-        let merged_schema = schema1.try_merge(schema2).unwrap();
-
-        let expected_schema = SchemaBuilder::new()
-            .measurement("my_measurement")
-            .tag("the_tag")
-            .build()
-            .unwrap();
-
-        assert_eq!(
-            expected_schema, merged_schema,
-            "\nExpected:\n{:#?}\nActual:\n{:#?}",
-            expected_schema, merged_schema
-        );
-    }
-
-    #[test]
-    fn test_merge_incompatible_schema_measurement_names() {
-        let schema1 = SchemaBuilder::new()
-            .tag("the_tag")
-            .measurement("measurement1")
-            .build()
-            .unwrap();
-
-        // different measurement name, same otherwise
-        let schema2 = SchemaBuilder::new()
-            .tag("the_tag")
-            .measurement("measurement2")
-            .build()
-            .unwrap();
-
-        let merged_schema_error = schema1.try_merge(schema2).unwrap_err();
-
-        assert_eq!(
-            merged_schema_error.to_string(),
-            "Schema Merge Error: Incompatible measurement names. Existing measurement name 'measurement1', new measurement name 'measurement2'"
-        );
-    }
-
-    #[test]
-    fn test_merge_incompatible_data_types() {
-        // same field name with different type
-        let schema1 = SchemaBuilder::new()
-            .field("the_field", ArrowDataType::Int16)
-            .build()
-            .unwrap();
-
-        // same field name with different type
-        let schema2 = SchemaBuilder::new()
-            .field("the_field", ArrowDataType::Int8)
-            .build()
-            .unwrap();
-
-        let merged_schema_error = schema1.try_merge(schema2).unwrap_err();
-
-        assert_eq!(merged_schema_error.to_string(), "Schema Merge Error: Incompatible data type for 'the_field'. Existing type Int16, new type Int8");
-    }
-
-    #[test]
-    fn test_merge_incompatible_column_types() {
-        let schema1 = SchemaBuilder::new().tag("the_tag").build().unwrap();
-
-        // same field name with different type
-        let schema2 = SchemaBuilder::new()
-            .influx_field("the_tag", Integer)
-            .build()
-            .unwrap();
-
-        let merged_schema_error = schema1.try_merge(schema2).unwrap_err();
-
-        assert_eq!(merged_schema_error.to_string(), "Schema Merge Error: Incompatible column type for 'the_tag'. Existing type Some(Field(Integer)), new type Some(Tag)");
-    }
-
-    #[test]
-    fn test_merge_incompatible_schema_nullability() {
-        let schema1 = SchemaBuilder::new()
-            .non_null_field("int_field", ArrowDataType::Int64)
-            .build()
-            .unwrap();
-
-        // same field name with different nullability
-        let schema2 = SchemaBuilder::new()
-            .field("int_field", ArrowDataType::Int64)
-            .build()
-            .unwrap();
-
-        let merged_schema_error = schema1.try_merge(schema2).unwrap_err();
-
-        assert_eq!(merged_schema_error.to_string(), "Schema Merge Error: Incompatible nullability for 'int_field'. Existing field can not be null, new field can be null");
     }
 
     #[test]
