+++ query/src/test.rs
@@ -18,10 +18,7 @@ use crate::{
 use crate::{exec::Executor, pruning::Prunable};
 
 use internal_types::{
-    schema::{
-        builder::{SchemaBuilder, SchemaMerger},
-        Schema,
-    },
+    schema::{builder::SchemaBuilder, merge::SchemaMerger, Schema},
     selection::Selection,
 };
 
@@ -228,14 +225,15 @@ impl TestChunk {
         }
         self.table_name = Some(table_name);
 
-        let mut merger = SchemaMerger::new().merge(new_column_schema).unwrap();
+        let mut merger = SchemaMerger::new();
+        merger.merge(&new_column_schema).unwrap();
 
-        if let Some(existing_schema) = self.table_schema.take() {
-            merger = merger
+        if let Some(existing_schema) = self.table_schema.as_ref() {
+            merger
                 .merge(existing_schema)
                 .expect("merging was successful");
         }
-        let new_schema = merger.build().unwrap();
+        let new_schema = merger.build();
 
         self.table_schema = Some(new_schema);
         self
